+++ src/components/searchfield/searchfield.js
@@ -252,7 +252,7 @@ SearchField.prototype = {
     }
 
     // Prevent browser typahead
-    this.element.attr('autocomplete', 'off');
+    this.element.attr('autocomplete', 'false');
 
     // Setup ARIA
     let label = this.element.attr('placeholder') || this.element.prev('label, .label').text().trim();