+++ src/components/spinbox/spinbox.js
@@ -574,7 +574,7 @@ Spinbox.prototype = {
 
     val = this.checkForNumeric(val);
     this.element[0].setAttribute('aria-valuenow', val || '0');
-    this.element[0].setAttribute('autocomplete', 'off');
+    this.element[0].setAttribute('autocomplete', 'false');
 
     // Toggle min/max buttons
     this.setIsDisabled(this.buttons.up, (val !== '' && max && val >= max) ? 'disable' : 'enable');